+++ jdbc/src/main/java/org/apache/metamodel/jdbc/JdbcDataContext.java
@@ -44,6 +44,7 @@ import org.apache.metamodel.data.DataSet;
 import org.apache.metamodel.data.EmptyDataSet;
 import org.apache.metamodel.data.MaxRowsDataSet;
 import org.apache.metamodel.data.ScalarFunctionDataSet;
+import org.apache.metamodel.jdbc.JdbcUtils.JdbcActionType;
 import org.apache.metamodel.jdbc.dialects.DB2QueryRewriter;
 import org.apache.metamodel.jdbc.dialects.DefaultQueryRewriter;
 import org.apache.metamodel.jdbc.dialects.H2QueryRewriter;
@@ -76,9 +77,12 @@ public class JdbcDataContext extends AbstractDataContext implements UpdateableDa
     public static final String SYSTEM_PROPERTY_BATCH_UPDATES = "metamodel.jdbc.batch.updates";
     public static final String SYSTEM_PROPERTY_CONVERT_LOBS = "metamodel.jdbc.convert.lobs";
 
-    public static final String SYSTEM_PROPERTY_COMPILED_QUERY_POOL_MAX_SIZE = "metamodel.jdbc.compiledquery.pool.max.size";
-    public static final String SYSTEM_PROPERTY_COMPILED_QUERY_POOL_MIN_EVICTABLE_IDLE_TIME_MILLIS = "metamodel.jdbc.compiledquery.pool.idle.timeout";
-    public static final String SYSTEM_PROPERTY_COMPILED_QUERY_POOL_TIME_BETWEEN_EVICTION_RUNS_MILLIS = "metamodel.jdbc.compiledquery.pool.eviction.period.millis";
+    public static final String SYSTEM_PROPERTY_COMPILED_QUERY_POOL_MAX_SIZE =
+            "metamodel.jdbc.compiledquery.pool.max.size";
+    public static final String SYSTEM_PROPERTY_COMPILED_QUERY_POOL_MIN_EVICTABLE_IDLE_TIME_MILLIS =
+            "metamodel.jdbc.compiledquery.pool.idle.timeout";
+    public static final String SYSTEM_PROPERTY_COMPILED_QUERY_POOL_TIME_BETWEEN_EVICTION_RUNS_MILLIS =
+            "metamodel.jdbc.compiledquery.pool.eviction.period.millis";
 
     public static final String DATABASE_PRODUCT_POSTGRESQL = "PostgreSQL";
     public static final String DATABASE_PRODUCT_MYSQL = "MySQL";
@@ -91,10 +95,10 @@ public class JdbcDataContext extends AbstractDataContext implements UpdateableDa
     public static final String DATABASE_PRODUCT_HIVE = "Apache Hive";
     public static final String DATABASE_PRODUCT_SQLITE = "SQLite";
 
-    public static final ColumnType COLUMN_TYPE_CLOB_AS_STRING = new ColumnTypeImpl("CLOB", SuperColumnType.LITERAL_TYPE,
-            String.class, true);
-    public static final ColumnType COLUMN_TYPE_BLOB_AS_BYTES = new ColumnTypeImpl("BLOB", SuperColumnType.BINARY_TYPE,
-            byte[].class, true);
+    public static final ColumnType COLUMN_TYPE_CLOB_AS_STRING =
+            new ColumnTypeImpl("CLOB", SuperColumnType.LITERAL_TYPE, String.class, true);
+    public static final ColumnType COLUMN_TYPE_BLOB_AS_BYTES =
+            new ColumnTypeImpl("BLOB", SuperColumnType.BINARY_TYPE, byte[].class, true);
 
     private static final Logger logger = LoggerFactory.getLogger(JdbcDataContext.class);
 
@@ -115,10 +119,9 @@ public class JdbcDataContext extends AbstractDataContext implements UpdateableDa
     private final String _databaseVersion;
 
     /**
-     * There are some confusion as to the definition of catalogs and schemas.
-     * Some databases seperate "groups of tables" by using schemas, others by
-     * catalogs. This variable indicates whether a MetaModel schema really
-     * represents a catalog.
+     * There are some confusion as to the definition of catalogs and schemas. Some databases seperate "groups of tables"
+     * by using schemas, others by catalogs. This variable indicates whether a MetaModel schema really represents a
+     * catalog.
      */
     private final String _identifierQuoteString;
     private final boolean _supportsBatchUpdates;
@@ -126,45 +129,33 @@ public class JdbcDataContext extends AbstractDataContext implements UpdateableDa
     private final boolean _usesCatalogsAsSchemas;
 
     /**
-     * Creates the strategy based on a data source, some table types and an
-     * optional catalogName
+     * Creates the strategy based on a data source, some table types and an optional catalogName
      * 
-     * @param dataSource
-     *            the datasource objcet to use for making connections
-     * @param tableTypes
-     *            the types of tables to include
-     * @param catalogName
-     *            a catalog name to use, can be null
+     * @param dataSource the datasource objcet to use for making connections
+     * @param tableTypes the types of tables to include
+     * @param catalogName a catalog name to use, can be null
      */
     public JdbcDataContext(DataSource dataSource, TableType[] tableTypes, String catalogName) {
         this(dataSource, null, tableTypes, catalogName);
     }
 
     /**
-     * Creates the strategy based on a {@link Connection}, some table types and
-     * an optional catalogName
+     * Creates the strategy based on a {@link Connection}, some table types and an optional catalogName
      * 
-     * @param connection
-     *            the database connection
-     * @param tableTypes
-     *            the types of tables to include
-     * @param catalogName
-     *            a catalog name to use, can be null
+     * @param connection the database connection
+     * @param tableTypes the types of tables to include
+     * @param catalogName a catalog name to use, can be null
      */
     public JdbcDataContext(Connection connection, TableType[] tableTypes, String catalogName) {
         this(null, connection, tableTypes, catalogName);
     }
 
     /**
-     * Creates the strategy based on a {@link DataSource}, some table types and
-     * an optional catalogName
+     * Creates the strategy based on a {@link DataSource}, some table types and an optional catalogName
      * 
-     * @param dataSource
-     *            the data source
-     * @param tableTypes
-     *            the types of tables to include
-     * @param catalogName
-     *            a catalog name to use, can be null
+     * @param dataSource the data source
+     * @param tableTypes the types of tables to include
+     * @param catalogName a catalog name to use, can be null
      */
     private JdbcDataContext(DataSource dataSource, Connection connection, TableType[] tableTypes, String catalogName) {
         _dataSource = dataSource;
@@ -192,7 +183,7 @@ public class JdbcDataContext extends AbstractDataContext implements UpdateableDa
         try {
             _isDefaultAutoCommit = con.getAutoCommit();
         } catch (SQLException e) {
-            throw JdbcUtils.wrapException(e, "determine auto-commit behaviour");
+            throw JdbcUtils.wrapException(e, "determine auto-commit behaviour", JdbcActionType.OTHER);
         }
 
         try {
@@ -256,8 +247,7 @@ public class JdbcDataContext extends AbstractDataContext implements UpdateableDa
     /**
      * Creates the strategy based on a {@link Connection}
      * 
-     * @param connection
-     *            the database connection
+     * @param connection the database connection
      */
     public JdbcDataContext(Connection connection) {
         this(connection, TableType.DEFAULT_TABLE_TYPES, null);
@@ -266,8 +256,7 @@ public class JdbcDataContext extends AbstractDataContext implements UpdateableDa
     /**
      * Creates the strategy based on a {@link DataSource}
      * 
-     * @param dataSource
-     *            the data source
+     * @param dataSource the data source
      */
     public JdbcDataContext(DataSource dataSource) {
         this(dataSource, TableType.DEFAULT_TABLE_TYPES, null);
@@ -298,7 +287,7 @@ public class JdbcDataContext extends AbstractDataContext implements UpdateableDa
                 result = false;
             }
         } catch (SQLException e) {
-            throw JdbcUtils.wrapException(e, "retrieve schema and catalog metadata");
+            throw JdbcUtils.wrapException(e, "retrieve schema and catalog metadata", JdbcActionType.METADATA);
         } finally {
             close(null);
         }
@@ -327,12 +316,13 @@ public class JdbcDataContext extends AbstractDataContext implements UpdateableDa
         final JdbcCompiledQueryLease lease = jdbcCompiledQuery.borrowLease();
         final DataSet dataSet;
         try {
-            dataSet = execute(lease.getConnection(), query, lease.getStatement(), jdbcCompiledQuery, lease, values);
+            dataSet =
+                    execute(lease.getConnection(), query, lease.getStatement(), jdbcCompiledQuery, lease, values, true);
         } catch (SQLException e) {
             // only close in case of an error - the JdbcDataSet will close
             // otherwise
             jdbcCompiledQuery.returnLease(lease);
-            throw JdbcUtils.wrapException(e, "execute compiled query");
+            throw JdbcUtils.wrapException(e, "execute compiled query", JdbcActionType.QUERY);
         } catch (RuntimeException e) {
             // only close in case of an error - the JdbcDataSet will close
             // otherwise
@@ -345,7 +335,8 @@ public class JdbcDataContext extends AbstractDataContext implements UpdateableDa
 
     @SuppressWarnings("resource")
     private DataSet execute(Connection connection, Query query, Statement statement, JdbcCompiledQuery compiledQuery,
-            JdbcCompiledQueryLease lease, Object[] values) throws SQLException, MetaModelException {
+            JdbcCompiledQueryLease lease, Object[] values, boolean closeConnectionOnCloseDataSet)
+            throws SQLException, MetaModelException {
         Integer maxRows = query.getMaxRows();
 
         final List<SelectItem> selectItems = query.getSelectClause().getItems();
@@ -473,7 +464,7 @@ public class JdbcDataContext extends AbstractDataContext implements UpdateableDa
             }
 
             if (lease == null) {
-                dataSet = new JdbcDataSet(query, this, connection, statement, resultSet);
+                dataSet = new JdbcDataSet(query, this, connection, statement, resultSet, closeConnectionOnCloseDataSet);
             } else {
                 dataSet = new JdbcDataSet(compiledQuery, lease, resultSet);
             }
@@ -497,23 +488,26 @@ public class JdbcDataContext extends AbstractDataContext implements UpdateableDa
     }
 
     public DataSet executeQuery(Query query) throws MetaModelException {
-
         final Connection connection = getConnection();
+        return executeQuery(connection, query, true);
+    }
+
+    protected DataSet executeQuery(Connection connection, Query query, boolean closeConnectionOnCloseDataSet) {
         final Statement statement;
         try {
             statement = connection.createStatement(ResultSet.TYPE_FORWARD_ONLY, ResultSet.CONCUR_READ_ONLY);
         } catch (SQLException e) {
-            throw JdbcUtils.wrapException(e, "create statement for query");
+            throw JdbcUtils.wrapException(e, "create statement for query", JdbcActionType.OTHER);
         }
 
         final DataSet dataSet;
         try {
-            dataSet = execute(connection, query, statement, null, null, null);
+            dataSet = execute(connection, query, statement, null, null, null, closeConnectionOnCloseDataSet);
         } catch (SQLException e) {
             // only close in case of an error - the JdbcDataSet will close
             // otherwise
             close(connection);
-            throw JdbcUtils.wrapException(e, "execute query");
+            throw JdbcUtils.wrapException(e, "execute query", JdbcActionType.QUERY);
         } catch (RuntimeException e) {
             // only close in case of an error - the JdbcDataSet will close
             // otherwise
@@ -539,8 +533,9 @@ public class JdbcDataContext extends AbstractDataContext implements UpdateableDa
 
     /**
      * Quietly closes the connection
-     *  @param connection The connection to close (if it makes sense, @see closeIfNecessary)
-
+     * 
+     * @param connection The connection to close (if it makes sense, @see closeIfNecessary)
+     * 
      */
     public void close(Connection connection) {
         closeIfNecessary(connection);
@@ -555,10 +550,8 @@ public class JdbcDataContext extends AbstractDataContext implements UpdateableDa
         FileHelper.safeClose(rs, st);
     }
 
-
     /**
-     * Convenience method to get the available catalogNames using this
-     * connection.
+     * Convenience method to get the available catalogNames using this connection.
      * 
      * @return a String-array with the names of the available catalogs.
      */
@@ -571,7 +564,7 @@ public class JdbcDataContext extends AbstractDataContext implements UpdateableDa
         try {
             metaData = connection.getMetaData();
         } catch (SQLException e) {
-            throw JdbcUtils.wrapException(e, "retrieve metadata");
+            throw JdbcUtils.wrapException(e, "retrieve metadata", JdbcActionType.METADATA);
         }
 
         // Retrieve catalogs
@@ -594,11 +587,10 @@ public class JdbcDataContext extends AbstractDataContext implements UpdateableDa
     }
 
     /**
-     * Gets the delegate from the JDBC API (ie. Connection or DataSource) that
-     * is being used to perform database interactions.
+     * Gets the delegate from the JDBC API (ie. Connection or DataSource) that is being used to perform database
+     * interactions.
      * 
-     * @return either a DataSource or a Connection, depending on the
-     *         configuration of the DataContext.
+     * @return either a DataSource or a Connection, depending on the configuration of the DataContext.
      */
     public Object getDelegate() {
         if (_dataSource == null) {
@@ -608,11 +600,11 @@ public class JdbcDataContext extends AbstractDataContext implements UpdateableDa
     }
 
     /**
-     * Gets an appropriate connection object to use - either a dedicated
-     * connection or a new connection from the datasource object.
+     * Gets an appropriate connection object to use - either a dedicated connection or a new connection from the
+     * datasource object.
      * 
-     * Hint: Use the {@link #close(Connection)} method to
-     * close the connection (and any ResultSet or Statements involved).
+     * Hint: Use the {@link #close(Connection)} method to close the connection (and any ResultSet or Statements
+     * involved).
      */
     public Connection getConnection() {
         if (_dataSource == null) {
@@ -621,7 +613,7 @@ public class JdbcDataContext extends AbstractDataContext implements UpdateableDa
         try {
             return _dataSource.getConnection();
         } catch (SQLException e) {
-            throw JdbcUtils.wrapException(e, "establish connection");
+            throw JdbcUtils.wrapException(e, "establish connection", JdbcActionType.OTHER);
         }
     }
 
@@ -710,7 +702,7 @@ public class JdbcDataContext extends AbstractDataContext implements UpdateableDa
                 }
 
             } catch (SQLException e) {
-                throw JdbcUtils.wrapException(e, "determine default schema name");
+                throw JdbcUtils.wrapException(e, "determine default schema name", JdbcActionType.METADATA);
             } finally {
                 closeIfNecessary(connection);
             }
@@ -750,8 +742,8 @@ public class JdbcDataContext extends AbstractDataContext implements UpdateableDa
     }
 
     /**
-     * Microsoft SQL Server returns users instead of schemas when calling
-     * metadata.getSchemas() This is a simple workaround.
+     * Microsoft SQL Server returns users instead of schemas when calling metadata.getSchemas() This is a simple
+     * workaround.
      * 
      * @return
      * @throws SQLException
@@ -819,7 +811,7 @@ public class JdbcDataContext extends AbstractDataContext implements UpdateableDa
             }
             return result;
         } catch (SQLException e) {
-            throw JdbcUtils.wrapException(e, "get schema names");
+            throw JdbcUtils.wrapException(e, "get schema names", JdbcActionType.METADATA);
         } finally {
             closeIfNecessary(connection);
         }
@@ -841,8 +833,14 @@ public class JdbcDataContext extends AbstractDataContext implements UpdateableDa
         return _fetchSizeCalculator;
     }
 
+    /**
+     * {@inheritDoc}
+     * 
+     * @throws RolledBackUpdateException when a potentially retriable {@link SQLException} was thrown during the
+     *             execution of the update script.
+     */
     @Override
-    public UpdateSummary executeUpdate(final UpdateScript update) {
+    public UpdateSummary executeUpdate(final UpdateScript update) throws RolledBackUpdateException {
         final JdbcUpdateCallback updateCallback;
 
         if (_supportsBatchUpdates && update instanceof BatchUpdateScript) {
@@ -863,11 +861,14 @@ public class JdbcDataContext extends AbstractDataContext implements UpdateableDa
                 update.run(updateCallback);
             }
             updateCallback.close(true);
+        } catch (UncheckedSQLException e) {
+            updateCallback.close(false);
+            throw new RolledBackUpdateException(e.getCause());
         } catch (RuntimeException e) {
             updateCallback.close(false);
             throw e;
         }
-        
+
         return updateCallback.getUpdateSummary();
     }